requirements.txt
aiogram==3.7.0
python-dotenv==1.0.1
aiosqlite==0.20.0
aiohttp==3.9.5
pydantic==2.7.1
# Babel==2.15.0 # Using simple JSON i18n for now, uncomment if switching
yookassa==3.5.0
pycountry==23.12.11
apscheduler==3.10.4
pydantic_settings